refactor: Removed deprecated CborSerializationUtil and used from comman module
