+++ plutus/src/main/java/com/bloxbean/cardano/client/plutus/spec/CostMdls.java
@@ -2,10 +2,10 @@ package com.bloxbean.cardano.client.plutus.spec;
 
 import co.nstant.in.cbor.CborException;
 import co.nstant.in.cbor.model.*;
+import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
 import com.bloxbean.cardano.client.exception.CborDeserializationException;
 import com.bloxbean.cardano.client.exception.CborRuntimeException;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
-import com.bloxbean.cardano.client.transaction.util.CborSerializationUtil;
 
 import java.util.HashMap;
