+++ plutus/src/main/java/com/bloxbean/cardano/client/plutus/util/ScriptDataHashGenerator.java
@@ -2,11 +2,11 @@ package com.bloxbean.cardano.client.plutus.util;
 
 import co.nstant.in.cbor.CborException;
 import co.nstant.in.cbor.model.Array;
+import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
 import com.bloxbean.cardano.client.crypto.Blake2bUtil;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
 import com.bloxbean.cardano.client.plutus.spec.PlutusData;
 import com.bloxbean.cardano.client.plutus.spec.Redeemer;
-import com.bloxbean.cardano.client.transaction.util.CborSerializationUtil;
 import com.bloxbean.cardano.client.util.HexUtil;
 
 import java.util.List;
