+++ plutus/src/test/java/com/bloxbean/cardano/client/plutus/spec/ListPlutusDataTest.java
@@ -3,9 +3,9 @@ package com.bloxbean.cardano.client.plutus.spec;
 import co.nstant.in.cbor.CborDecoder;
 import co.nstant.in.cbor.CborException;
 import co.nstant.in.cbor.model.DataItem;
+import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
 import com.bloxbean.cardano.client.exception.CborDeserializationException;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
-import com.bloxbean.cardano.client.transaction.util.CborSerializationUtil;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigInteger;
