+++ plutus/src/test/java/com/bloxbean/cardano/client/plutus/spec/serializers/MapDataJsonSerializerTest.java
@@ -1,11 +1,11 @@
 package com.bloxbean.cardano.client.plutus.spec.serializers;
 
 import co.nstant.in.cbor.model.Map;
+import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
 import com.bloxbean.cardano.client.plutus.spec.BigIntPlutusData;
 import com.bloxbean.cardano.client.plutus.spec.BytesPlutusData;
 import com.bloxbean.cardano.client.plutus.spec.ListPlutusData;
 import com.bloxbean.cardano.client.plutus.spec.MapPlutusData;
-import com.bloxbean.cardano.client.transaction.util.CborSerializationUtil;
 import com.bloxbean.cardano.client.util.HexUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
