+++ transaction-spec/src/main/java/com/bloxbean/cardano/client/transaction/spec/cert/GenesisKeyDelegation.java
@@ -10,7 +10,7 @@ import lombok.*;
 
 import java.util.List;
 
-import static com.bloxbean.cardano.client.transaction.util.CborSerializationUtil.toBytes;
+import static com.bloxbean.cardano.client.common.cbor.CborSerializationUtil.toBytes;
 
 @Getter
 @NoArgsConstructor
