+++ transaction-spec/src/main/java/com/bloxbean/cardano/client/transaction/spec/script/NativeScript.java
@@ -4,10 +4,10 @@ import co.nstant.in.cbor.CborException;
 import co.nstant.in.cbor.model.Array;
 import co.nstant.in.cbor.model.DataItem;
 import co.nstant.in.cbor.model.UnsignedInteger;
+import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
 import com.bloxbean.cardano.client.exception.CborDeserializationException;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
 import com.bloxbean.cardano.client.spec.Script;
-import com.bloxbean.cardano.client.transaction.util.CborSerializationUtil;
 import com.bloxbean.cardano.client.transaction.util.NativeScriptDeserializer;
 import com.bloxbean.cardano.client.util.JsonUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
