+++ transaction-spec/src/test/java/com/bloxbean/cardano/client/transaction/spec/cert/PoolRegistrationTest.java
@@ -1,10 +1,10 @@
 package com.bloxbean.cardano.client.transaction.spec.cert;
 
 import co.nstant.in.cbor.CborException;
+import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
 import com.bloxbean.cardano.client.exception.CborDeserializationException;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
 import com.bloxbean.cardano.client.spec.UnitInterval;
-import com.bloxbean.cardano.client.transaction.util.CborSerializationUtil;
 import com.bloxbean.cardano.client.util.HexUtil;
 import org.junit.jupiter.api.Test;
